fix issue with immporting utils
