+++ packages/gollm/common/utils.py
@@ -4,9 +4,6 @@ import re
 import tiktoken
 
 from common.LlmToolsInterface import LlmToolsInterface
-from llms.azure.AzureTools import AzureTools
-from llms.llama.LlamaTools import LlamaTools
-from llms.openai.OpenAiTools import OpenAiTools
 
 
 def validate_schema(schema):
@@ -261,17 +258,3 @@ def format_json_to_schema(schema, data):
         return format_object(schema, data)
     else:
         return format_value(schema, data)
-
-
-def determine_llm(model: str) -> LlmToolsInterface:
-    """
-    Determine the LLM model based on the provided model name.
-    """
-    if "llama" in model.lower():
-        return LlamaTools()
-    elif "openai" in model.lower():
-        return OpenAiTools()
-    elif "azure" in model.lower():
-        return AzureTools()
-    else:
-        raise ValueError(f"Unknown LLM model: {model}")
