+++ packages/gollm/tasks/latex_to_sympy.py
@@ -1,9 +1,9 @@
 import sys
 import traceback
 from chains import latex_to_sympy_chain
-from common.utils import determine_llm
 from entities import EquationsModel
 from llms.azure.AzureTools import AzureTools
+from llms.determine_llm import determine_llm
 
 from taskrunner import TaskRunnerInterface
